+++ .editorconfig
@@ -236,6 +236,3 @@ dotnet_diagnostic.SA1600.severity = none
 
 # SA1602: Enumeration items should be documented
 dotnet_diagnostic.SA1602.severity = none
-
-# SA1649: File name should match first type name
-dotnet_diagnostic.SA1649.severity = none
