+++ src/NHSD.BuyingCatalogue.Ordering.Domain/Results/Result{T}.cs
